% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populi.R
\docType{data}
\name{Employee_data}
\alias{Employee_data}
\title{Sample employee data}
\format{
A data frame with 50 rows and 4 variables
}
\usage{
data(Employee_data)
}
\description{
A data frame containing a randomly generated log of start and finish times for employees. The variables are as follows:
}
\details{
\itemize{
\item Employee_ID. Unique identifier of a employee
\item Start_time. The hour that the employee started work in 24 hour format
\item Finish_time. The hour that the employee finished work in 24 hour format
\item Hours_worked. The sum of hours that the employee worked that day
}
}
\keyword{datasets}
