% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populi.R
\docType{data}
\name{Patient_data}
\alias{Patient_data}
\title{Sample public hospital patient data}
\format{
A data frame with 1000 rows and 4 variables
}
\usage{
data(Patient_data)
}
\description{
A data frame containing randomly generated demographics of patients born between 01/01/2000 and 05/01/2023. The variables are as follows:
}
\details{
\itemize{
\item Patient_number. Unique identifier of a patient
\item Date_of_birth. Date of birth of the patient
\item Country_of_birth. The country the patient was born in
\item Needs_interpreter. Binary indicator of the patient's need for an interpreter (0, 1)
}
}
\keyword{datasets}
